+++ src/video/window.rs
@@ -134,9 +134,8 @@ impl<'video> Window<'video> {
     ///
     /// Returns `Err` if failed to make it full screen.
     pub fn full_screen(&self) -> Result<()> {
-        let ret = unsafe {
-            bind::SDL_SetWindowFullscreen(self.as_ptr(), bind::SDL_WINDOW_FULLSCREEN as u32)
-        };
+        let ret =
+            unsafe { bind::SDL_SetWindowFullscreen(self.as_ptr(), bind::SDL_WINDOW_FULLSCREEN) };
         if ret != 0 {
             return Err(crate::SdlError::Others { msg: Sdl::error() });
         }
